+++ src/app/pages/home/home.component.ts
@@ -3,8 +3,8 @@ import { CardDetails } from 'src/app/core/interfaces/card-details';
 import { CardProfessional } from 'src/app/core/interfaces/card-professional';
 import { CardVacancy } from 'src/app/core/interfaces/card-vacancy';
 import { CardDetailsService } from 'src/app/core/service/cardDetails/card-details.service';
-import { CardProfessionalService } from 'src/app/core/service/cardProfessional/card-professional.service';
-import { CardVacancyService } from 'src/app/core/service/cardVacancy/card-vacancy.service';
+import { CardProfessionalService } from 'src/app/core/service/professional/professional.service';
+import { CardVacancyService } from 'src/app/core/service/vacancy/vacancy.service';
 
 @Component({
   selector: 'app-home',
